Implement request ownership, isolation, and authorization

+++ app/Http/Controllers/ServiceRequestController.php
@@ -9,7 +9,9 @@ class ServiceRequestController extends Controller
 {
     public function index()
     {
-        $requests = ServiceRequest::latest()->get();
+        $requests = ServiceRequest:: where('user_id', auth()->id())
+        ->latest()
+        ->get();
         return view('requests.index', compact('requests'));
     }
 
@@ -22,25 +24,37 @@ class ServiceRequestController extends Controller
             'tracking_id'   => 'nullable|string|max:50',
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = auth()->id();
 
         ServiceRequest::create($data);
 
         return redirect()->route('requests.index');
     }
+    
     public function destroy(ServiceRequest $request)
     {
+        if ($request->user_id !== auth()->id()){
+            abort(403);
+        }
         $request->delete();
 
         return redirect()->route('requests.index');
     }
+
     public function edit(ServiceRequest $request)
     {
+        if ($request->user_id !== auth()->id()){
+            abort(403);
+        }
+
         return view('requests.edit', compact('request'));
     }
 
     public function update(Request $httpRequest, ServiceRequest $request)
     {
+        if ($request->user_id !== auth()->id()){
+            abort(403);
+        }
         $data = $httpRequest->validate([
             'service_type' => 'required|string|max:50',
             'sender_name' => 'required|string|max:150',
